Moved clean up signature test and added actions mod
