+++ integration-tests/tests/lib.rs
@@ -189,100 +189,6 @@ mod check {
     }
 }
 
-mod wait_for {
-    use crate::MultichainTestContext;
-    use anyhow::Context;
-    use backon::ExponentialBuilder;
-    use backon::Retryable;
-    use mpc_contract::ProtocolContractState;
-    use mpc_contract::RunningContractState;
-    use mpc_recovery_node::web::StateView;
-
-    pub async fn running_mpc<'a>(
-        ctx: &MultichainTestContext<'a>,
-        epoch: u64,
-    ) -> anyhow::Result<RunningContractState> {
-        let is_running = || async {
-            let state: ProtocolContractState = ctx
-                .rpc_client
-                .view(ctx.nodes.ctx().mpc_contract.id(), "state", ())
-                .await?;
-
-            match state {
-                ProtocolContractState::Running(running) if running.epoch >= epoch => Ok(running),
-                ProtocolContractState::Running(running) => {
-                    anyhow::bail!("running with an older epoch: {}", running.epoch)
-                }
-                _ => anyhow::bail!("not running"),
-            }
-        };
-        is_running
-            .retry(&ExponentialBuilder::default().with_max_times(6))
-            .await
-            .with_context(|| format!("mpc nodes did not reach epoch '{epoch}' before deadline"))
-    }
-
-    pub async fn has_at_least_triples<'a>(
-        ctx: &MultichainTestContext<'a>,
-        id: usize,
-        expected_triple_count: usize,
-    ) -> anyhow::Result<StateView> {
-        let is_enough_triples = || async {
-            let state_view: StateView = ctx
-                .http_client
-                .get(format!("{}/state", ctx.nodes.url(id)))
-                .send()
-                .await?
-                .json()
-                .await?;
-
-            match state_view {
-                StateView::Running { triple_count, .. }
-                    if triple_count >= expected_triple_count =>
-                {
-                    Ok(state_view)
-                }
-                StateView::Running { .. } => anyhow::bail!("node does not have enough triples yet"),
-                StateView::NotRunning => anyhow::bail!("node is not running"),
-            }
-        };
-        is_enough_triples
-            .retry(&ExponentialBuilder::default().with_max_times(6))
-            .await
-            .with_context(|| format!("mpc node '{id}' failed to generate '{expected_triple_count}' triples before deadline"))
-    }
-
-    pub async fn has_at_least_presignatures<'a>(
-        ctx: &MultichainTestContext<'a>,
-        id: usize,
-        expected_presignature_count: usize,
-    ) -> anyhow::Result<StateView> {
-        let is_enough_presignatures = || async {
-            let state_view: StateView = ctx
-                .http_client
-                .get(format!("{}/state", ctx.nodes.url(id)))
-                .send()
-                .await?
-                .json()
-                .await?;
-
-            match state_view {
-                StateView::Running {
-                    presignature_count, ..
-                } if presignature_count >= expected_presignature_count => Ok(state_view),
-                StateView::Running { .. } => {
-                    anyhow::bail!("node does not have enough presignatures yet")
-                }
-                StateView::NotRunning => anyhow::bail!("node is not running"),
-            }
-        };
-        is_enough_presignatures
-            .retry(&ExponentialBuilder::default().with_max_times(6))
-            .await
-            .with_context(|| format!("mpc node '{id}' failed to generate '{expected_presignature_count}' presignatures before deadline"))
-    }
-}
-
 trait MpcCheck {
     type Response;
